refactor(Note): make a different model for Notes
